Basics Update
I redid and added onto the program skeleton that Tyler created.
I decided to erase the window class and have the main class hold
multiple state classes to switch between. Each state class will have a
render and an update class. (I want to make classes that build off of
the gamestate class, but I do not know how to do that in C++ yet...) I
added a level and tile class for the gamestate to hold.

The tile class holds two numbers, the # for the location of the sprite
texture on the sprite map, and the z value (altitude).

The level class holds a 2d dynamic array of tiles called the map. The
level (will) receive a pointer/reference (not sure which) and will draw
the level by taking tileimg#'s from the tiles and grabbing the textures
off of the level spritesheet.

Our level files are held under Resources/Visual/Karl (will be changed to
company name)/Test (level names)/...

WARNING: Currently, this commit has an error. I believe this is due to
my setup being wrong, but since I can not fix the error, I am leaving it
up to y'all to fix it. Sorry.