Controls Overhaul update
This update completely redid the control scheme to allow for update key
layouts later on and saving information about certain keys. Also the
frame work for battling is introduced, but a whole lot more work has to
be done to the battling framework and drawing fonts to the screen before
we can continue on from this point.

Not this update visually performs 90% like the last update, but loads of
things under the hood have been revamped.

New Control Schemes

Classic Control (the only one implemented)

'L' acts as the 'A' button (select button)
'J' acts as the 'B' button (back button)

Mouse Control Scheme (not implemented yet)

We will also allow users to point and click at stuff.