Just a few edits  to the control system
This is not an important commit, but I have tweaked some controls.
Changes:
-Added a start and select button (num pad 3 and 1), and renamed the old
select button to accept button ('a' for accept) The new accept and back
buttons are on numpad6 and numpad4 respectively. I also changed the
height of the player sprite and messed around with his speed and added a
delay to his start walking time.

TODO:
-Player animations
-Get some good player & NPC sprites
-Work on the battling
-Work on loading and saving text files
-Work on main menu