Fix for the directory system
PLEASE READ THIS CAVEAT
In order for this update to work, you must go to "Variables.h" and set
your path to the resources folder there, then save your version of the
path in the comment above.

I also changed the order the player is drawn in so that he is displayed
properly in front of an NPC.

EDIT: This commit was edited January 25, 2016 to remove files that we don't have the right to distribute
Original commit was uploaded on January 24th.