A Few Updates
Created a text class to render our text, I have been working for 5 and a
half hours trying different methods and we are sticking with the one we
have now, because it is great!

I also fixed the directory, full addresses weren't required, TL;DR: you
don't need to edit anything in "Variables.h" anymore.

FINAL TL;DR
We have fonts!

To do:
-Loading levels from files (cleaning up Level class)
-Get better fonts
-Create NPC spritesheet
-Create player animations (time % numAnimations)
-Battle System and Javamon information
-Save files