Basic File Loading and an Attempt at Cleaning
I cleaned up the level class (a little, but changes start small) and
created a very basic level loading system.
Also, every std::string was converted to std::wstring or sf::String.
This is to allow for unicode character encodings.

TODO:
-Move the dialogue stuff into the Player class (in preparations for
servers) and create a class that Player and NPC both extend (to make
things cleaner).
-Make the Player that exists in the Level.cpp an array and make the
level that exist in the Main.cpp an array also for multiplayer
compatibility in the future
-Clean up the classes and add comments
-Add destructors when appropriate
-Not hold the language file in the level class, that should be moved to
the player class (or convert the controls class to a subsection of a
general settings class that is held by the player class through a level
method (therefore, add a level.addPlayer method))
-Convert arrays to whatever the C++ equivalent of Java's ArrayLists
(arrays with non-static lengths that you can add or remove variables
with ease)

Comment structure
/*
(Description)

Author:
Date Created: (not needed for previously written functions)
Date Revised:
Last Revised by:
*/

I hope I mentioned everything, but I am prone to have left many things
out it in the changes and the to do lists.