Intermediate Overhaul Update
This is a section of the new overhaul update that I am introducing.
Things are being moved around and cleaned up so that my code will start
to look presentable.

Several new classes, the screen, entity, and loading class

All menus will be based off the screen class and all dynamic items on
the map will be based off of the Entity class. Also loading WILL be
moved to the KFile class (strip down the StringEditor class and
upgrading the loading system in the Level class)

Work that needs to be done:
-Entity class
-Battle stuff
-Menu class
-Level class (text file loading, code cleanup)
-Player class (animations)
-Multithreading