Something happened
I did a massive overhaul to the layout of the program. Multithreading
still looks like an absolute pain to mess with. Other than a loss in the
number of commands read by the Level.cpp, no functionality really has
been lost or gained (other than I changed certain variables to vectors,
which will eliminate the need for a setNumberOf NPCs or Teleports
command)
TLDR: Layout change for the better mostly