31st Commit
I'm counting the initial as the first commit.

I released a to do list today on the drive, so I figured I might as well
release my current build so that we are all on the same basis.

Nothing special has been done, but you can open up a cyan coloured menu
if you press NumPad3 while walking around in the level (doesn't do
anything)

If you have any suggestions, please feel free to leave a comment on the
file or send a message in the group chat. (You should get Skype,
Clay...)