Workin' on the in Game Menu
I have started working on the in-game menu. There are a few things to
note.

The in-game menu class is held inside of the player class. This seems
like the way to go for multiplayer purposes, but because I do not know
how to use C++, I do not know how to make both the player and in-game
menu class include each other so I can pass the player as a variable in
the InGameMenu::render and InGameMenu::update methods.
The other thing to note is a bad design flaw that I do not now how to
fix yet. In order to quit to main menu: the in-game menu class alerts
the level class (through the Player::getInGameMenu), the level class
alerts the single player class, and the single player class alerts the
main class. How am I going to do this set-up for multiplayer (since the
level reports it wants to return to main menu, it doesn't mention who
wants to return to the main menu.)

Because I expect us to change the name of our game soon, I made the
color scheme gray and blue instead of the nasty orange and yellow
colours I was using earlier.